projects/presentation_analyzer/requirements.txt
# Google Drive API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Data processing
pandas==2.1.4
numpy==1.25.2

# Database
sqlite3

# Configuration
PyYAML==6.0.1

# Utilities
python-dateutil==2.8.2
requests==2.31.0
pathlib2==2.3.7

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

